Made table derivation tool more flexible
So apparently my RAM is broken. That took me a while to figure out.
While I get a replacement, I have to restrict myself to a weaker
test for the intermediate tables: instead of deriving them from
the huge table and checking that they are the same, I have to derive
a small h0k2 table from the intermediate ones and check that it is
correct. This is not a 100% proof of correctness, but it is good
enough (and much faster).