Removed slow test
This test generated a pruning table and it was too slow (around a minute).
Now the whole test suite runs in 49s on my laptop, which is more reasonable.

Pruning table generation should be tested with tools. They run in release
mode and are much faster, but can be still run in debug mode optionally.